wip: reorganize propagator.test.cpp's tests
