+++ CMakeLists.txt
@@ -393,7 +393,7 @@ IF (BUILD_UNIT_TESTS)
 
     ENABLE_TESTING ()
 
-    FILE (GLOB_RECURSE TEST_SRCS "${PROJECT_SOURCE_DIR}/test/${PROJECT_PATH}/*.test.cpp")
+    FILE (GLOB_RECURSE TEST_SRCS "${PROJECT_SOURCE_DIR}/test/${PROJECT_PATH}/*.test.cpp" "${PROJECT_SOURCE_DIR}/test/${PROJECT_PATH}/*.validation.cpp")
 
     ADD_EXECUTABLE (${UNIT_TESTS_TARGET} ${TEST_SRCS})
 
@@ -463,7 +463,7 @@ IF (BUILD_VALIDATION_TESTS)
 
     ENABLE_TESTING ()
 
-    FILE (GLOB_RECURSE TEST_SRCS "${PROJECT_SOURCE_DIR}/validation/${PROJECT_PATH}/*.cpp")
+    FILE (GLOB_RECURSE TEST_SRCS "${PROJECT_SOURCE_DIR}/validation/${PROJECT_PATH}/*.cpp" "${PROJECT_SOURCE_DIR}/validation/${PROJECT_PATH}/*.test.cpp" "${PROJECT_SOURCE_DIR}/validation/${PROJECT_PATH}/*.validation.cpp")
 
     ADD_EXECUTABLE (${VALIDATION_TESTS_TARGET} ${TEST_SRCS})
