--- validation/OpenSpaceToolkit/Astrodynamics/data/scenarios/scenario005-mission-sequence.yaml
@@ -1,48 +1,0 @@
-type: MISSION_SEQUENCE
-data:
-  spacecraft:
-    mass: 100.0
-    drag-cross-section: 1.0
-    drag-coefficient: 2.2
-    orbit:
-      type: CARTESIAN
-      data:
-        date:
-          time-scale: UTC
-          value: "2023-01-01T00:00:00"
-        frame: GCRF
-        body: EARTH
-        x: -4283387.412456233
-        y: -4451426.776125101
-        z: -2967617.850750065
-        vx: 4948.074939732174
-        vy: -957.3429532772124
-        vz: -5721.173027553034
-
-  sequence:
-    max-duration: 86400.1
-    forces:
-      - type: GRAVITY
-        data:
-          body: EARTH
-          model: EGM96
-          degree: 0
-          order: 0
-    propagator:
-      type: RUNGE_KUTTA_4
-      data:
-        step: 30.0
-    segments:
-      - type: COAST
-        data:
-          stop-condition:
-            type: RELATIVE_TIME
-            data:
-              duration: 86400.0
-
-  output:
-    step: 120.0
-    include:
-      - ELAPSED_SECONDS
-      - CARTESIAN_POSITION_GCRF
-      - CARTESIAN_VELOCITY_GCRF
